requirements.txt
watchdog
PyMuPDF
python-docx
textract
